backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & Auth
supabase==2.3.0
psycopg2-binary==2.9.9

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Payment Processing
stripe==7.6.0

# Voice AI
elevenlabs==1.0.0

# Communication
sendgrid==6.10.0
twilio==8.10.0

# Google Calendar
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
aiofiles==23.2.1

# Date/Time
python-dateutil==2.8.2

# CORS
python-multipart==0.0.6